Parameterise O diffusion TMD tests by compound and restructure outputs by model
