+++ ml_peg/analysis/nebs/O_diffusion_2D_TMDs/metrics.yml
@@ -4,34 +4,34 @@ metrics:
     bad: 0.4
     unit: eV
     tooltip: "Diffusion barrier error for MoS2"
-    level_of_theory: DFT-PBE
+    level_of_theory: PBE
   MoSe2 barrier error:
     good: 0.05
     bad: 0.4
     unit: eV
     tooltip: "Diffusion barrier error for MoSe2"
-    level_of_theory: DFT-PBE
+    level_of_theory: PBE
   MoTe2 barrier error:
     good: 0.05
     bad: 0.4
     unit: eV
     tooltip: "Diffusion barrier error for MoTe2"
-    level_of_theory: DFT-PBE
+    level_of_theory: PBE
   WS2 barrier error:
     good: 0.05
     bad: 0.4
     unit: eV
     tooltip: "Diffusion barrier error for WS2"
-    level_of_theory: DFT-PBE
+    level_of_theory: PBE
   WSe2 barrier error:
     good: 0.05
     bad: 0.4
     unit: eV
     tooltip: "Diffusion barrier error for WSe2"
-    level_of_theory: DFT-PBE
+    level_of_theory: PBE
   WTe2 barrier error:
     good: 0.05
     bad: 0.4
     unit: eV
     tooltip: "Diffusion barrier error for WTe2"
-    level_of_theory: DFT-PBE
+    level_of_theory: PBE
