+++ ml_peg/calcs/nebs/O_diffusion_2D_TMDs/calc_O_diffusion_2D_TMDs.py
@@ -11,8 +11,8 @@ from janus_core.calculations.neb import NEB
 import pytest
 
 from ml_peg.calcs.utils.utils import download_s3_data
+from ml_peg.models import current_models
 from ml_peg.models.get_models import load_models
-from ml_peg.models.models import current_models
 
 MODELS = load_models(current_models)
 
@@ -47,14 +47,14 @@ def relaxed_structs() -> dict[str, Atoms]:
             for state in ["initial", "end"]:
                 struct_name = f"{compound}_{state}.xyz"
                 struct = read(structure_dir / struct_name)
-                struct.calc = calc.get_calculator()
+                struct.calc = calc.get_calculator(precision="high")
                 struct.info.setdefault("charge", 0)
                 struct.info.setdefault("spin", 1)
 
                 geomopt = GeomOpt(
                     struct=struct,
                     write_results=True,
-                    file_prefix=OUT_PATH / f"{struct_name}-{model_name}",
+                    file_prefix=OUT_PATH / model_name / struct_name,
                     filter_class=None,
                 )
                 geomopt.run()
@@ -63,10 +63,13 @@ def relaxed_structs() -> dict[str, Atoms]:
 
 
 @pytest.mark.slow
+@pytest.mark.parametrize("compound", COMPOUNDS)
 @pytest.mark.parametrize("model_name", MODELS)
-def test_o_diffusion_mos2(relaxed_structs: dict[str, Atoms], model_name: str) -> None:
+def test_o_diffusion(
+    relaxed_structs: dict[str, Atoms], model_name: str, compound: str
+) -> None:
     """
-    Run calculations required for oxygen adatom diffusion on MoS2.
+    Run calculations required for oxygen adatom diffusion on a TMD compound.
 
     Parameters
     ----------
@@ -74,199 +77,26 @@ def test_o_diffusion_mos2(relaxed_structs: dict[str, Atoms], model_name: str) ->
         Relaxed input structures, indexed by structure name and model name.
     model_name
         Name of model to use.
+    compound
+        Name of compound to use.
     """
     try:
         NEB(
-            init_struct=relaxed_structs[f"MoS2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"MoS2_end.xyz-{model_name}"],
+            init_struct=relaxed_structs[f"{compound}_initial.xyz-{model_name}"],
+            final_struct=relaxed_structs[f"{compound}_end.xyz-{model_name}"],
             n_images=11,
             interpolator="pymatgen",
             minimize=True,
             write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_MoS2-{model_name}",
+            file_prefix=OUT_PATH / model_name / f"O_diffusion_{compound}",
         ).run()
     except Exception:
         NEB(
-            init_struct=relaxed_structs[f"MoS2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"MoS2_end.xyz-{model_name}"],
+            init_struct=relaxed_structs[f"{compound}_initial.xyz-{model_name}"],
+            final_struct=relaxed_structs[f"{compound}_end.xyz-{model_name}"],
             n_images=11,
             interpolator="ase",
             minimize=True,
             write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_MoS2-{model_name}",
-        ).run()
-
-
-@pytest.mark.slow
-@pytest.mark.parametrize("model_name", MODELS)
-def test_o_diffusion_mose2(relaxed_structs: dict[str, Atoms], model_name: str) -> None:
-    """
-    Run calculations required for oxygen adatom diffusion on MoSe2.
-
-    Parameters
-    ----------
-    relaxed_structs
-        Relaxed input structures, indexed by structure name and model name.
-    model_name
-        Name of model to use.
-    """
-    try:
-        NEB(
-            init_struct=relaxed_structs[f"MoSe2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"MoSe2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="pymatgen",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_MoSe2-{model_name}",
-        ).run()
-    except Exception:
-        NEB(
-            init_struct=relaxed_structs[f"MoSe2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"MoSe2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="ase",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_MoSe2-{model_name}",
-        ).run()
-
-
-@pytest.mark.slow
-@pytest.mark.parametrize("model_name", MODELS)
-def test_o_diffusion_mote2(relaxed_structs: dict[str, Atoms], model_name: str) -> None:
-    """
-    Run calculations required for oxygen adatom diffusion on MoTe2.
-
-    Parameters
-    ----------
-    relaxed_structs
-        Relaxed input structures, indexed by structure name and model name.
-    model_name
-        Name of model to use.
-    """
-    try:
-        NEB(
-            init_struct=relaxed_structs[f"MoTe2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"MoTe2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="pymatgen",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_MoTe2-{model_name}",
-        ).run()
-    except Exception:
-        NEB(
-            init_struct=relaxed_structs[f"MoTe2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"MoTe2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="ase",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_MoTe2-{model_name}",
-        ).run()
-
-
-@pytest.mark.slow
-@pytest.mark.parametrize("model_name", MODELS)
-def test_o_diffusion_ws2(relaxed_structs: dict[str, Atoms], model_name: str) -> None:
-    """
-    Run calculations required for oxygen adatom diffusion on WS2.
-
-    Parameters
-    ----------
-    relaxed_structs
-        Relaxed input structures, indexed by structure name and model name.
-    model_name
-        Name of model to use.
-    """
-    try:
-        NEB(
-            init_struct=relaxed_structs[f"WS2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"WS2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="pymatgen",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_WS2-{model_name}",
-        ).run()
-    except Exception:
-        NEB(
-            init_struct=relaxed_structs[f"WS2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"WS2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="ase",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_WS2-{model_name}",
-        ).run()
-
-
-@pytest.mark.slow
-@pytest.mark.parametrize("model_name", MODELS)
-def test_o_diffusion_wse2(relaxed_structs: dict[str, Atoms], model_name: str) -> None:
-    """
-    Run calculations required for oxygen adatom diffusion on WSe2.
-
-    Parameters
-    ----------
-    relaxed_structs
-        Relaxed input structures, indexed by structure name and model name.
-    model_name
-        Name of model to use.
-    """
-    try:
-        NEB(
-            init_struct=relaxed_structs[f"WSe2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"WSe2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="pymatgen",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_WSe2-{model_name}",
-        ).run()
-    except Exception:
-        NEB(
-            init_struct=relaxed_structs[f"WSe2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"WSe2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="ase",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_WSe2-{model_name}",
-        ).run()
-
-
-@pytest.mark.slow
-@pytest.mark.parametrize("model_name", MODELS)
-def test_o_diffusion_wte2(relaxed_structs: dict[str, Atoms], model_name: str) -> None:
-    """
-    Run calculations required for oxygen adatom diffusion on WTe2.
-
-    Parameters
-    ----------
-    relaxed_structs
-        Relaxed input structures, indexed by structure name and model name.
-    model_name
-        Name of model to use.
-    """
-    try:
-        NEB(
-            init_struct=relaxed_structs[f"WTe2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"WTe2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="pymatgen",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_WTe2-{model_name}",
-        ).run()
-    except Exception:
-        NEB(
-            init_struct=relaxed_structs[f"WTe2_initial.xyz-{model_name}"],
-            final_struct=relaxed_structs[f"WTe2_end.xyz-{model_name}"],
-            n_images=11,
-            interpolator="ase",
-            minimize=True,
-            write_band=True,
-            file_prefix=OUT_PATH / f"O_diffusion_WTe2-{model_name}",
+            file_prefix=OUT_PATH / model_name / f"O_diffusion_{compound}",
         ).run()
